requirements.txt
flask
groq
opencv-python-headless
numpy
Pillow
python-dotenv
werkzeug